sns-auto-poster/requirements.txt
anthropic>=0.40.0
pytrends>=4.9.2
feedparser>=6.0.11
schedule>=1.2.2
flask>=3.1.0
python-dotenv>=1.0.1
requests>=2.32.3
